refactor:minor code changes
